requirement.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.4.0
awscli==1.22.82
blis==0.7.5
botocore==1.24.27
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
cryptography==36.0.1
cymem==2.0.6
distlib==0.3.4
docutils==0.15.2
docx2txt==0.8
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
fastapi==0.73.0
filelock==3.4.2
Flask==2.0.0
flatbuffers==2.0
gast==0.5.3
google-api-core==1.31.5
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-cloud-vision==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.55.0
googletrans==4.0.0rc1
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
langcodes==3.3.0
libclang==13.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
murmurhash==1.0.6
nltk==3.6.7
numpy==1.22.1
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
pathy==0.6.1
pdf2image==1.16.0
pdfminer.six==20211012
Pillow==9.0.1
platformdirs==2.4.1
preshed==3.0.6
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.13.0
pydantic==1.8.2
pyparsing==3.0.7
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyresparser==1.0.6
pyrsistent==0.18.1
pytesseract==0.3.9
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.5
pytz==2021.3
PyWavelets==1.3.0
PyYAML==5.4.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.5.2
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.0.13
tifffile==2022.3.16
torch==1.11.0
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.1
virtualenv==20.13.0
wasabi==0.9.0
Werkzeug==2.0.2
wrapt==1.13.3
